let ptlslog optional (enabled by default on Linux and macOS)
