+++ t/cli.c
@@ -49,6 +49,7 @@
 #endif
 #include "picotls.h"
 #include "picotls/openssl.h"
+#include "picotls/ptlslog.h"
 #if PICOTLS_USE_BROTLI
 #include "picotls/certificate_compression.h"
 #endif
