+++ t/picotls.c
@@ -31,6 +31,9 @@
 #include "picotls/pembase64.h"
 #include "../deps/picotest/picotest.h"
 #include "../lib/picotls.c"
+#if PICOTLS_USE_PTLSLOG
+#include "../lib/ptlslog.c"
+#endif
 #include "test.h"
 
 static void test_is_ipaddr(void)
@@ -1691,6 +1694,8 @@ static void test_tls12_hello(void)
 
 static void test_escape_json_unsafe_string(void)
 {
+#if PICOTLS_USE_PTLSLOG
+
 #define STRLIT(s) s, sizeof(s) - 1
     char buf[100];
     size_t escaped_len;
@@ -1704,6 +1709,8 @@ static void test_escape_json_unsafe_string(void)
 
     escaped_len = escape_json_unsafe_string(buf, STRLIT("\x00 \x1f \x7f"));
     ok(strcmp(buf, "\\u0000 \\u001f \\u007f") == 0);
+
+#endif /* PICOTLS_USE_PTLSLOG */
 }
 
 void test_picotls(void)
